Allow to set the signature method like this:
require 'oauth/signature/plaintext'

consumer = OAuth::Consumer.new(
...
  :signature_method => 'PLAINTEXT'
)

I hope it was intended to work like that. The option on the consumer
was previosly named :oauth_signature_method but I couldn't find
anything in the code that relied on it.

It also fixes the plaintext signature encoding, assuming that the
implementation at term.ie is correct.

This patch was originally mailed to the oauth-ruby list on 15 May 2008