allow a redirect with different host, but same path
When the response returns a redirect with a location that has a different host, but the path is still the same an error was thrown.
In the case of a change of host however, we do want to continue with a new request. The site and http object have to be updated with the new host url and a new request can be made.